Merged PR 5478: Steward API
This is a full port of the Steward V2 API modified to work with AAD auth. In addition to porting everything over, I've also reworked the file structure into something sensible, aligned all namespaces to be logical, revisited endpoint paths per David's request.

What remains is verifying a working build in Azure, porting tests, and getting a release pipeline setup. That is the purview of my next task.