Merged PR 5795: Fix prod release pipeline tests.
Use the aud claim instead of the appid claim.
The prod app registration was not providing the AppId claim type. I did a little poking and noticed that the dev app registration didn't have an access token accepted version defined. I updated it to '2' to match prod, and then neither returned the AppId claim type. Instead the claim type to use is AUD, which is defined by Azure as, "In id_tokens, the audience is your app's Application ID, assigned to your app in the Azure portal." That's exactly what we want.