Merged PR 5911: LiveOpsGiftHistory only records successful bans
LiveOpsGiftHistory was recording all attempted bans, which I don't think is the behavior we want. This change ensures that failed bans are not recorded. I will need help from someone with access to the prod tables in gde003 to help me remove the invalid xuid entries that have made it into Kusto in the last 30 days to get the failing GetGiftHistory_InvalidXuid tests that pointed out this issue passing once more.