Merged PR 6013: Remove Replace and Reset inventory endpoints
This change removes the Replace and Reset endpoints. We don't use them in production, and in testing them we slow down future lookups of the profile. This is why Gravity tests eventually start timing out.
Along with this change I made a fresh Gravity account and put it's details into our integration tests. This should be the last time I ever have to do that.