Merged PR 7378: Steward UI: Clear route reuse cache on app switch & set default user-details back to FH4
Switching apps sometimes causes random data loss which could put routes in some bad states (ex: no player is selected but the tool remembers the player reference to take action on).

Clearing out the cache on app change is the simplest way to not run into this issue until we find a way to actually fix it.