Merged PR 7313: No longer upload Agent email to our BackgroundJobs, GiftHistory, or BanHistory
This change modifies the behavior of our logging. Now instead of banhistory/gifthistory/jobs storing the agents email field they will store their unique Azure Object ID. In addition to this, every time such a query is run we update a Cosmos table to keep an up to date record of the email and name to associate with the Object ID.

Also included in this PR is a change to background jobs. Specifically I have updated the model so that the Rowkey is now the JobId instead of the PartitionKey. The PartitionKey is now the calling User's Object ID.

This change does not include a mass lookup of jobs by UserId, or the addition of a Viewed column for BackgroundJobs, those will come in following PR.