Merged PR 7687: Add Steelhead APIs
This change integrates all the new Steelhead APIs into Steward. I have tested all of the services endpoints and they are working as expected.

For the moment all Steelhead APIs require a LiveOpsAdmin role to query. This is due to two things. 1: Not all needed Kusto data has made it to gde003 so any of the inventory naming/id logic will be able to retrieve the reference tables. 2: Steelhead doesn't have a stable retail endpoint yet, all testing was done on an endpoint deployed for us to test on. We will remove this restriction once relevant Kusto data arrives and the retail endpoint is stood up.