Merged PR 8059: Reuse ForzaClient
Prior to this change, every request to the LSP generated a new ForzaClient and Service. From discussion with Services, we should be reusing the ForzaClient, and there's no issue with creating a new service for each request.

As such, for each title I've combined all the service wrappers into a single service wrapper, and each wrapper now has a single forzaclient used by all requests, instead of making a new one for each request.