Merged PR 8024: Changing requestingAgent to requestorObjectId
Now that we actually track jobs by Azure object ID, update the naming of variable, contracts, and kusto data to reflect as much. The key to this working is in the updated KustoQueries that coalesce the requestingAgent and requestorObjectId columns from the *History tables, allowing us to safely switch while still picking up the correct requestingAgent value from entries that predate the requestorObjectId's creation.