Merged PR 8310: Sunrise Profile Notes
Based on Luke's old branch, I took it and got it up to date with all the sunrise API changes that've happened.
Adds functionality to add and look at profile notes for a user.
Had to change our invalid xuid for testing, due to a bug in LSP call that created a user if the one we queried for didn't exist.