Merged PR 9730: Fix Cypress testing on Staging/Prod (by disabling SRI on the fly)
Ripped the snippet at the top of the page from here:
https://github.com/cypress-io/cypress/issues/2393#issuecomment-465377636

This has to be the first script tag on the page and only triggers if the site is being operated from Cypress.

When additional script nodes are added, it adds a failure handler to the node. When this failure handler is triggered, it checks if it is possibly due to SRI and replaces the script with a non-SRI-enabled version.

Normally I might say "this is a security concern" but:
- you can already automate a browser in a fashion where scripts being SRI doesn't matter (thru Selenium and friends)
- this only takes effect when operating the site thru Cypress, so it won't affect "normal" users
    - it could also be triggered from an extension or other web plugin with full site access, but those scripts could already do anything - so this doesn't affect the attack surface in those cases.