Merged PR 9820: Steward UI: Fix ExpressionChangedAfterItHasBeenCheckedError on player selection tool + BONUS WIN (づ◕‿‿◕)づ
**What I found:**
- was only occurring on player-selection-single (not bulk)
- the expression changed was mat-focus from 'true' -> 'false'
- the problem was the input being disabled when knownIdentities were found (this.disable = knownidentities.length > 0)

**Solution:**
The purpose of disabling the input is to stop additional identities from being processed when there is an identity already active in the single player selection chip list.

Rather than disabling, we can use the [hidden] attribute to hide the input. Mat-focus is no longer changed and we no longer see the error in console.

**Bonus Win!!!**
The space between the identity chip and bottom line is gone now (originally was the disabled input)

**Before**
![image.png](https://dev.azure.com/turn10/4043f02f-acc8-4003-8330-9187c43a2ab0/_apis/git/repositories/99f68dc4-7f8d-4b91-9173-efd898e50254/pullRequests/9820/attachments/image.png)

**After**
![image (2).png](https://dev.azure.com/turn10/4043f02f-acc8-4003-8330-9187c43a2ab0/_apis/git/repositories/99f68dc4-7f8d-4b91-9173-efd898e50254/pullRequests/9820/attachments/image%20%282%29.png)