Merged PR 9682: Complete working V1 endpoint selection.
Big PR, but basic changes.
Added concept of endpoint to Apollo, Sunrise, Woodstock, and Steelhead. Now any API for those titles that interacts with the LSP will interact with prod by default, but allows passing of an endpointKey header which can be used to select other endpoints to run the request against.

From the UI side of this, endpoint selection is stored in local storage and defaults to 'Production' for Sunrise/Apollo and 'Development' for Steelhead and Woodstock due to them not yet having production instances. When not on Production, the title selection navbars will display additional information informing the support agents that they're acting against a different LSP endpoint than expected. Endpoint selection can be changed via the dropdowns in the settings flyout.