Merged PR 10231: Modify Client to query API for list of valid endpoints to select.
Prior to this change StewardAPI and StewardClient maintained separate lists of supported endpoints. This change makes the StewardAPI the source for both, and includes a new state to track the state of endpoint key lists. It the event the list of endpoints isn't available or a user is forced off of an endpoint selection because it no longer exists we raise a snackbar to warn/inform the user.