Merged PR 11134: Make all Error Spinners based on a size of 1em; spinner demo
Errors spinners should be size 1em so they play nicely with inline text.
Angular's default `mat-spinner` has its size manually set in pixels, presumably for compatibility with older browsers.
We do not care about that, so all spinners should be based on a default 1em font-size as is normal for other libraries.
For larger/smaller error spinners, change the font-size of the error spinner.

## Before
![image (3).png](https://dev.azure.com/turn10/4043f02f-acc8-4003-8330-9187c43a2ab0/_apis/git/repositories/99f68dc4-7f8d-4b91-9173-efd898e50254/pullRequests/11134/attachments/image%20%283%29.png)

## After
![image (2).png](https://dev.azure.com/turn10/4043f02f-acc8-4003-8330-9187c43a2ab0/_apis/git/repositories/99f68dc4-7f8d-4b91-9173-efd898e50254/pullRequests/11134/attachments/image%20%282%29.png)