Merged PR 11132: Add studio endpoints for Live Ops Admins and tweak endpoint selection margins.
Live ops admins will be able to see and select the studio endpoints, but it will not exist for any other user roles.
With switching from codenames to title abbreviations we scrunched up the endpoint selection displays in router link buttons, this change gives them more room to be readable.