Merged PR 11166: Remove 'workaround' Kusto queries for endpoint selection.
When adding the endpoint column to already existing Kusto tables, we had to special case the query for entries that existed prior to the endpoint column's existence. Now that 30+ days have passed and all pre-endpoint selection entries have expired we no longer need the workaround queries and they can be removed.