Merged PR 12105: Integrate ServiceFactories into ServiceWrappers and remove duplicate code
Refactored Sunrise and Woodstock ServiceWrappers to utilize Sunrise and Woodstock ServiceFactory classes. Also removed direct references to Factory classes from providers for the time being, and added in unit tests for the two methods that were untestable prior. Likely Apollo and Opus to follow, I think the Factory stuff being pulled into it's own class is very clean.

https://dev.azure.com/t10motorsport/Motorsport/_workitems/edit/917275
with a couple hours of investigation on testability per this task (basic results in discussion)
https://dev.azure.com/t10motorsport/Motorsport/_workitems/edit/917265