Merged PR 12281: Reduce jasmine deprecation warnings; Make prettier ignore .angular/
We pin all versions of `jasmine-core` to 3.3.0 in order to avoid deprecation warnings and other issues. Some nested dependencies were requiring `jasmine-core` @ `^3.6.0` which was causing issues.

The wall of v3 to v4 deprecation warnings were added in `3.99.0`. We should disable this version pinning when we intend to migrate to prepare for the migration.