Merged PR 14818: Add new wrapper for loading component
I think the goal of this was to move common functionality to a helper/wrapper. I'm not 100% happy with this current version because of the random <ng-template> that seemingly does nothing but is needed for the wrapper to load the information of the component once the monitor is successful.

<ng-content> can't work as far as I know because it cannot "produce" content but only projects already existing content.

*UPDATE*
Changed it from a component to a directive which I looks way more clean.