Merged PR 17973: Fix localized string dropdown
Bug was created by me with motd ui. I needed a non-required localized string dropdown so I removed the required validator in the subcomponent for localized string to let the parent component deal with it.

Reverted changing the {} to null for the default of the formcontrols. And now the required validator is only removed in the component when specified.