Merged PR 18753: Fh5 ban history comparator now properly handles multiple live ops entries with matching IDs.
Because we now correlate by ID, and ban extensions use the same ID, it's possible to get more than a single live ops ban entry that matches the service entries.

To account for this, we will always prefer the entry with the longest expire time, to accurately represent the most recent state of the ban we know.