Merged PR 19481: Fix reauth/retry workflow being cancelled when refreshing access token
GET LspEndpoints is one of the first Steward endpoints we call.
If that fails with a 401, it follows our reauth/retry which refreshes the user's access token.

The problem is that after refreshing the access token, we automatically try to get and reverify lsp endpoints.
This request to GET LspEndpoints cancels the original request, ultimately cancelling the reauth/retry workflow.