Merged PR 24519: Change the way we get key vault values
Implemented a class similar to what is in forza.net.
This is a work in progress to validate the change before I change it everywhere
SteelheadPegasusService is an example of what the new way to get Key Vault values would be
KeyVaultConfig has 2 ways of getting values. One is how Steward does it usually and the other on is from forza.net repo. Not sure if we also wanted to change the way we query the keys.
Probably worth a small technical park